Add a script to generate SerializationMacros.hpp
SerializationMacros.hpp has been re-generated using this script, with
the same number of macros as before. The script regularizes the
generated macros a bit in order to make them easier to generate, so the
"new" SerializationMacros.hpp appears to have changed a lot, but it
should be functionally the same.